\name{Cheat Function}
\alias{cheat}
\title{Cheat }
\description{You can use the cheat function by entering the question number into the function. This works for Questions Q3.1.2, Q3.1.6, Q3.1.9m Q3.1.17 of the PIPS course.
}
\usage{
  cheat(exercise)
}

\arguments{
\item{exercise}{The exercise for cheating}
}

